fix settings preservation, and switch to less-ugly tri tiles.
Store and persist all settings in attributes on the main <window/>, rather than the Settings dialog, as the latter meant they persisted even if the user pressed Cancel.

Also, add a "show grid lines" option, which still needs some graphics and css changes to work for square and hex tiles.